requirements.txt
Mesa-Geo==0.5.0
Mesa==1.1.1
pillow==10.2.0
pandas==2.2.0
numpy~=1.26.4
rasterio==1.3.9
shapely==2.0.3
geopandas==0.14.3